refactoring Card and renaming to ValueCard
