+++ sparkle/src/components/ValueCard.tsx
@@ -1,35 +1,16 @@
-import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
 
-import { Spinner } from "@sparkle/components";
+import { CardButton, Spinner } from "@sparkle/components";
 import { cn } from "@sparkle/lib/utils";
 
-const CARD_SIZES = ["xs", "sm", "md", "lg"] as const;
-export type CardSizeType = (typeof CARD_SIZES)[number];
-
-const cardVariants = cva(
-  "s-flex s-flex-col s-gap-2 s-rounded-2xl s-bg-structure-50 s-p-4 s-min-h-[128px]",
-  {
-    variants: {
-      size: {
-        xs: "s-w-[180px]",
-        sm: "s-w-[240px]",
-        md: "s-w-[300px]",
-        lg: "s-w-[360px]",
-      },
-    },
-    defaultVariants: {
-      size: "sm",
-    },
-  }
-);
-
-interface CardRootProps extends VariantProps<typeof cardVariants> {
+interface CardRootProps {
   children: React.ReactNode;
   className?: string;
 }
-const Root = ({ size, className, children }: CardRootProps) => (
-  <div className={cn(cardVariants({ size }), className)}>{children}</div>
+const Root = ({ className, children }: CardRootProps) => (
+  <CardButton size="md" className={cn("s-flex s-flex-col s-gap-2", className)}>
+    {children}
+  </CardButton>
 );
 
 interface CardHeaderProps {
@@ -47,9 +28,7 @@ interface CardTitleProps {
 }
 
 const Title = ({ className, children }: CardTitleProps) => (
-  <div
-    className={cn("s-text-sm s-font-semibold s-text-element-800", className)}
-  >
+  <div className={cn("s-text-sm s-font-semibold s-text-foreground", className)}>
     {children}
   </div>
 );
@@ -60,7 +39,7 @@ interface CardSubtitleProps {
 }
 
 const Subtitle = ({ className, children }: CardSubtitleProps) => (
-  <div className={cn("s-text-sm s-text-element-700", className)}>
+  <div className={cn("s-text-sm s-text-muted-foreground", className)}>
     {children}
   </div>
 );
@@ -79,12 +58,16 @@ const Content = ({
   if (isLoading) {
     return (
       <div className="s-flex s-items-center s-justify-start">
-        <Spinner size="sm" />
+        <Spinner size="xs" variant="dark" />
       </div>
     );
   }
   return (
-    <div className={cn("s-flex s-flex-col s-gap-3", className)}>{children}</div>
+    <div
+      className={cn("s-flex s-flex-col s-gap-3 s-text-foreground", className)}
+    >
+      {children}
+    </div>
   );
 };
 
@@ -105,21 +88,18 @@ interface CardProps {
   content: React.ReactNode;
   footer?: React.ReactNode;
   isLoading?: boolean;
-  size?: CardSizeType;
   className?: string;
 }
 
-export const Card = ({
+export const ValueCard = ({
   title,
   subtitle,
   content,
   footer,
   isLoading = false,
-  size = "sm",
-  className,
 }: CardProps) => {
   return (
-    <Root size={size} className={className}>
+    <Root>
       <Header>
         <Title>{title}</Title>
         {subtitle && <Subtitle>{subtitle}</Subtitle>}
